created autController to implement registration user logic and start develop for role service logic

+++ src/main/java/org/farfenix/service/UserServiceImpl.java
@@ -41,7 +41,10 @@ public class UserServiceImpl implements IUserService{
     @Transactional
     @Override
     public void deleteUser(String email) {
-        userRepository.deleteByEmaik(email);
+        User theUser = getUser(email);
+        if (theUser != null) {
+            userRepository.deleteByEmaik( email);
+        }
     }
 
     @Override
